requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cloudevents==1.2.0
deprecation==2.1.0
env-validate==0.4
Flask==2.1.1
functions-framework==3.0.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
neo4j==4.4.2
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
tweepy==4.8.0
urllib3==1.26.9
watchdog==1.0.2
Werkzeug==2.1.1
zipp==3.8.0
